requirements.txt
# Core dependencies
fastapi
uvicorn
sqlalchemy
mysql-connector-python
pymongo
pandas
scikit-learn
spacy
nltk
pydantic
python-multipart
python-dotenv
pytest
sib-api-v3-sdk
